requirements.txt
langchain
langchain-community
langchain-ollama
pillow
langchain-google-genai